#12 Pull out FewValues and ManyValuesThresholdFinder strategy classes. Will now do further refactoring to reduce code duplication. Tests pass. WIP.
