+++ src/main/scala/org/apache/spark/mllib/feature/MDLPDiscretizer.scala
@@ -17,8 +17,6 @@
 
 package org.apache.spark.mllib.feature
 
-import scala.collection.mutable
-
 import org.apache.spark.storage.StorageLevel
 import org.apache.spark.Logging
 import org.apache.spark.rdd._
@@ -96,217 +94,7 @@ class MDLPDiscretizer private (val data: RDD[LabeledPoint],
       }             
     })
   }
-  
-  /**
-   * Evaluate boundary points and select the most relevant. This version is used when 
-   * the number of candidates exceeds the maximum size per partition (distributed version).
-   * 
-   * @param candidates RDD of candidates points (point, class histogram).
-   * @param maxBins Maximum number of points to select
-   * @param elementsByPart Maximum number of elements to evaluate in each partition.
-   * @return Sequence of threshold values.
-   */
-  private def getThresholds(
-      candidates: RDD[(Float, Array[Long])],
-      maxBins: Int, 
-      elementsByPart: Int): Seq[Float] = {
 
-    // Get the number of partitions according to the maximum size established by partition
-    val partitions = { x: Long => math.ceil(x.toFloat / elementsByPart).toInt }    
-    
-    // Insert the extreme values in the stack (recursive iteration)
-    val stack = new mutable.Queue[((Float, Float), Option[Float])]
-    stack.enqueue(((Float.NegativeInfinity, Float.PositiveInfinity), None))
-    var result = Seq(Float.NegativeInfinity) // # points = # bins - 1
-
-    while (stack.nonEmpty && result.size < maxBins){
-      val (bounds, lastThresh) = stack.dequeue
-      // Filter the candidates between the last limits added to the stack
-      var cands = candidates.filter({ case (th, _) => th > bounds._1 && th < bounds._2 })
-      val nCands = cands.count
-      if (nCands > 0) {
-        cands = cands.coalesce(partitions(nCands))
-        // Selects one threshold among the candidates and returns two partitions to recurse
-        evalThresholds(cands, lastThresh) match {
-          case Some(th) =>
-            result = th +: result
-            stack.enqueue(((bounds._1, th), Some(th)))
-            stack.enqueue(((th, bounds._2), Some(th)))
-          case None => /* criteria not fulfilled, finish! */
-        }
-      }
-    }
-
-    result.sorted :+ Float.PositiveInfinity
-  }
-  
-  /**
-   * Evaluates boundary points and selects the most relevant candidates (sequential version).
-   * Here, the evaluation is bounded by partition as the number of points is small enough.
-   * 
-   * @param candidates RDD of candidates points (point, class histogram).
-   * @param maxBins Maximum number of points to select.
-   * @return Sequence of threshold values.
-   */
-  private def getThresholds(candidates: Array[(Float, Array[Long])], maxBins: Int) = {
-
-    val stack = new mutable.Queue[((Float, Float), Option[Float])]
-    // Insert first in the stack (recursive iteration)
-    stack.enqueue(((Float.NegativeInfinity, Float.PositiveInfinity), None))
-    var result = Seq(Float.NegativeInfinity) // # points = # bins - 1
-
-    while (stack.nonEmpty && result.size < maxBins){
-      val (bounds, lastThresh) = stack.dequeue
-      // Filter the candidates between the last limits added to the stack
-      val newCandidates = candidates.filter({ case (th, _) => 
-          th > bounds._1 && th < bounds._2 
-        })      
-      if (newCandidates.length > 0) {
-        evalThresholds(newCandidates, lastThresh, nLabels) match {
-          case Some(th) =>
-            result = th +: result
-            stack.enqueue(((bounds._1, th), Some(th)))
-            stack.enqueue(((th, bounds._2), Some(th)))
-          case None => /* criteria not fulfilled, finish */
-        }
-      }
-    }
-    result.sorted :+ Float.PositiveInfinity
-  }
-
-  /**
-   * Compute entropy minimization for candidate points in a range,
-   * and select the best one according to the MDLP criterion (RDD version).
-   * 
-   * @param candidates RDD of candidate points (point, class histogram).
-   * @param lastSelected Last selected threshold.
-   * @return The minimum-entropy candidate.
-   */
-  private def evalThresholds(
-      candidates: RDD[(Float, Array[Long])],
-      lastSelected : Option[Float]) = {
-
-    val sc = candidates.sparkContext
-
-    // Compute the accumulated frequencies by partition
-    val totalsByPart = sc.runJob(candidates, { case it =>
-      val accum = Array.fill(nLabels)(0L)
-      for ((_, freqs) <- it) {for (i <- 0 until nLabels) accum(i) += freqs(i)}
-      accum
-    }: (Iterator[(Float, Array[Long])]) => Array[Long])
-    
-    // Compute the total frequency for all partitions
-    var totals = Array.fill(nLabels)(0L)
-    for (t <- totalsByPart) totals = (totals, t).zipped.map(_ + _)
-    val bcTotalsByPart = sc.broadcast(totalsByPart)
-    val bcTotals = sc.broadcast(totals)
-
-    val result = candidates.mapPartitionsWithIndex({ (slice, it) =>
-      // Accumulate frequencies from the left to the current partition
-      var leftTotal = Array.fill(nLabels)(0L)
-      for (i <- 0 until slice) leftTotal = (leftTotal, bcTotalsByPart.value(i)).zipped.map(_ + _)
-      var entropyFreqs = Seq.empty[(Float, Array[Long], Array[Long], Array[Long])]
-      // ... and from the current partition to the rightmost partition
-      for ((cand, freqs) <- it) {
-        leftTotal = (leftTotal, freqs).zipped.map(_ + _)
-        val rightTotal = (bcTotals.value, leftTotal).zipped.map(_ - _)
-        entropyFreqs = (cand, freqs, leftTotal.clone, rightTotal) +: entropyFreqs
-      }        
-      entropyFreqs.iterator
-    })
-
-    // calculate h(S)
-    // s: number of elements
-    // k: number of distinct classes
-    // hs: entropy        
-    val s  = totals.sum
-    val hs = entropy(totals.toSeq, s)
-    val k  = totals.count(_ != 0)
-
-    // select the best threshold according to MDLP
-    val finalCandidates = result.flatMap({
-      case (cand, _, leftFreqs, rightFreqs) =>
-        val k1 = leftFreqs.count(_ != 0)
-        val s1 = if (k1 > 0) leftFreqs.sum else 0
-        val hs1 = entropy(leftFreqs, s1)
-        val k2 = rightFreqs.count(_ != 0)
-        val s2 = if (k2 > 0) rightFreqs.sum else 0
-        val hs2 = entropy(rightFreqs, s2)
-        val weightedHs = (s1 * hs1 + s2 * hs2) / s
-        val gain = hs - weightedHs
-        val delta = log2(math.pow(3, k) - 2) - (k * hs - k1 * hs1 - k2 * hs2)
-        var criterion = (gain - (log2(s - 1) + delta) / s) > stoppingCriterion
-        lastSelected match {
-          case None =>
-          case Some(last) => criterion = criterion && (cand != last)
-        }
-        if (criterion) Seq((weightedHs, cand)) else Seq.empty[(Double, Float)]
-    })
-    // Select among the list of accepted candidate, that with the minimum weightedHs
-    if (finalCandidates.count > 0) Some(finalCandidates.min._2) else None
-  }
-  
-  /**
-   * Compute entropy minimization for candidate points in a range,
-   * and select the best one according to MDLP criterion (sequential version).
-   * 
-   * @param candidates Array of candidate points (point, class histogram).
-   * @param lastSelected last selected threshold.
-   * @param nLabels Number of classes.
-   * @return The minimum-entropy cut point.
-   * 
-   */
-  private def evalThresholds(
-      candidates: Array[(Float, Array[Long])],
-      lastSelected : Option[Float],
-      nLabels: Int): Option[Float] = {
-    
-    // Calculate the total frequencies by label
-    val totals = candidates.map(_._2).reduce((freq1, freq2) => (freq1, freq2).zipped.map(_ + _))
-    
-    // Compute the accumulated frequencies (both left and right) by label
-    var leftAccum = Array.fill(nLabels)(0L)
-    var entropyFreqs = Seq.empty[(Float, Array[Long], Array[Long], Array[Long])]
-    for(i <- candidates.indices) {
-      val (cand, freq) = candidates(i)
-      leftAccum = (leftAccum, freq).zipped.map(_ + _)
-      val rightTotal = (totals, leftAccum).zipped.map(_ - _)
-      entropyFreqs = (cand, freq, leftAccum.clone, rightTotal) +: entropyFreqs
-    }
-
-    // calculate h(S)
-    // s: number of elements
-    // k: number of distinct classes
-    // hs: entropy
-    val s = totals.sum
-    val hs = entropy(totals.toSeq, s)
-    val k = totals.count(_ != 0)
-
-    // select best threshold according to the criteria
-    val finalCandidates = entropyFreqs.flatMap({
-      case (cand, _, leftFreqs, rightFreqs) =>
-        val k1 = leftFreqs.count(_ != 0)
-        val s1 = if (k1 > 0) leftFreqs.sum else 0
-        val hs1 = entropy(leftFreqs, s1)
-        val k2 = rightFreqs.count(_ != 0)
-        val s2 = if (k2 > 0) rightFreqs.sum else 0
-        val hs2 = entropy(rightFreqs, s2)
-        val weightedHs = (s1 * hs1 + s2 * hs2) / s
-        val gain = hs - weightedHs
-        val delta = log2(math.pow(3, k) - 2) - (k * hs - k1 * hs1 - k2 * hs2)
-        var criterion = (gain - (log2(s - 1) + delta) / s) > stoppingCriterion
-        
-        lastSelected match {
-            case None =>
-            case Some(last) => criterion = criterion && (cand != last)
-        }
-
-        if (criterion) Seq((weightedHs, cand)) else Seq.empty[(Double, Float)]
-    })
-    // Select among the list of accepted candidate, that with the minimum weightedHs
-    if (finalCandidates.nonEmpty) Some(finalCandidates.min._2) else None
-  }
- 
   /**
    * Run the entropy minimization discretizer on input data.
    * 
@@ -363,19 +151,19 @@ class MDLPDiscretizer private (val data: RDD[LabeledPoint],
     }
     
     // Group elements by feature and point (get distinct points)
-    val nonzeros = featureValues.reduceByKey{ case (v1, v2) => 
+    val nonZeros = featureValues.reduceByKey{ case (v1, v2) =>
       (v1, v2).zipped.map(_ + _)
     }
       
     // Add zero elements for sparse data
-    val zeros = nonzeros
+    val zeros = nonZeros
       .map{case ((k, p), v) => (k, v)}
       .reduceByKey{ case (v1, v2) =>  (v1, v2).zipped.map(_ + _)}
       .map{ case (k, v) => 
         val v2 = for(i <- v.indices) yield bclassDistrib.value(i) - v(i)
         ((k, 0.0F), v2.toArray)
       }.filter{case (_, v) => v.sum > 0}
-    val distinctValues = nonzeros.union(zeros)
+    val distinctValues = nonZeros.union(zeros)
     
     // Sort these values to perform the boundary points evaluation
     val sortedValues = distinctValues.sortByKey()
@@ -403,19 +191,21 @@ class MDLPDiscretizer private (val data: RDD[LabeledPoint],
     val bBigIndexes = sc.broadcast(bigIndexes)
       
     // The features with a small number of points can be processed in a parallel way
+    val smallThresholdsFinder = new FewValuesThresholdFinder(nLabels, stoppingCriterion)
     val smallThresholds = initialCandidates
       .filter{case (k, _) => !bBigIndexes.value.contains(k) }
       .groupByKey()
       .mapValues(_.toArray)
-      .mapValues(points => getThresholds(points.toArray.sortBy(_._1), maxBins))
+      .mapValues(points => smallThresholdsFinder.findThresholds(points.sortBy(_._1), maxBins))
     
     // Feature with too many points must be processed iteratively (rare condition)
     logInfo("Number of features that exceed the maximum size per partition: " + 
         bigIndexes.size)
     var bigThresholds = Map.empty[Int, Seq[Float]]
+    val bigThresholdsFinder = new ManyValuesThresholdFinder(nLabels, stoppingCriterion)
     for (k <- bigIndexes.keys){ 
       val cands = initialCandidates.filter{case (k2, _) => k == k2}.values.sortByKey()
-      bigThresholds += ((k, getThresholds(cands, maxBins, elementsByPart)))
+      bigThresholds += ((k, bigThresholdsFinder.findThresholds(cands, maxBins, elementsByPart)))
     }
 
     // Join all thresholds in a single structure
@@ -439,25 +229,8 @@ class MDLPDiscretizer private (val data: RDD[LabeledPoint],
 /** Companion object for static members */
 object MDLPDiscretizer {
 
-  private val LOG2 = math.log(2)
-
-  /** @return log base 2 of x */
-  private val log2 = { x: Double => math.log(x) / LOG2 }
-
   /** The original paper suggested 0 for the stopping criterion, but smaller values like -1e-3 yield more splits */
   private val DEFAULT_STOPPING_CRITERION = 0
-
-  /**
-    * @param freqs sequence of integer frequencies.
-    * @param n the sum of all the frequencies in the list.
-    * @return the total entropy
-    */
-  private def entropy(freqs: Seq[Long], n: Long) = {
-    -freqs.aggregate(0.0)(
-      { case (h, q) => h + (if (q == 0) 0 else (q.toDouble / n) * log2(q.toDouble / n))},
-      { case (h1, h2) => h1 + h2 }
-    )
-  }
 
   /** @return true if f1 and f2 define a boundary */
   private val isBoundary = (f1: Array[Long], f2: Array[Long]) => {
